refactored SecureSocket to use interface #4313

+++ src/lib/base/String.cpp
@@ -171,7 +171,7 @@ findReplaceAll(
 String
 removeFileExt(String filename)
 {
-	unsigned dot = filename.find_last_of('.');
+	size_t dot = filename.find_last_of('.');
 
 	if (dot == String::npos) {
 		return filename;
